+++ lib/extensions/zip.py
@@ -72,31 +72,50 @@ class Zip(InkstitchExtension):
             stitch_plan = self._make_offsets(stitch_plan)
 
         base_file_name = self._get_file_name()
-        path = tempfile.mkdtemp()
+        path = None
+        files = []
+        temp_file_name = None
 
-        files = self.generate_output_files(stitch_plan, path, base_file_name)
+        try:
+            path = tempfile.mkdtemp()
 
-        if not files:
-            errormsg(_("No embroidery file formats selected."))
+            files = self.generate_output_files(stitch_plan, path, base_file_name)
 
-        temp_file = tempfile.NamedTemporaryFile(suffix=".zip", delete=False)
+            if not files:
+                errormsg(_("No embroidery file formats selected."))
 
-        # in windows, failure to close here will keep the file locked
-        temp_file.close()
+            temp_file = tempfile.NamedTemporaryFile(suffix=".zip", delete=False)
+            temp_file_name = temp_file.name
 
-        with ZipFile(temp_file.name, "w") as zip_file:
+            # in windows, failure to close here will keep the file locked
+            temp_file.close()
+
+            with ZipFile(temp_file_name, "w") as zip_file:
+                for file in files:
+                    zip_file.write(file, os.path.basename(file))
+
+            # inkscape will read the file contents from stdout and copy
+            # to the destination file that the user chose
+            with open(temp_file_name, 'rb') as output_file:
+                sys.stdout.buffer.write(output_file.read())
+        finally:
+            if temp_file_name is not None:
+                try:
+                    os.remove(temp_file_name)
+                except OSError:
+                    pass
+
             for file in files:
-                zip_file.write(file, os.path.basename(file))
+                try:
+                    os.remove(file)
+                except OSError:
+                    pass
 
-        # inkscape will read the file contents from stdout and copy
-        # to the destination file that the user chose
-        with open(temp_file.name, 'rb') as output_file:
-            sys.stdout.buffer.write(output_file.read())
-
-        os.remove(temp_file.name)
-        for file in files:
-            os.remove(file)
-        os.rmdir(path)
+            if path is not None:
+                try:
+                    os.rmdir(path)
+                except OSError:
+                    pass
 
         # don't let inkex output the SVG!
         sys.exit(0)
